Types changed to stdint

+++ sw/demo/demo.c
@@ -1,4 +1,5 @@
 #include <msp430.h>
+#include <stdint.h>
 
 #include "platform.h"
 #include "clock.h"
@@ -15,7 +16,7 @@
 #define STATE_YELLOW    4
 #define STATE_WHITE     5
 
-void sleep(unsigned int sleepcnt);
+void sleep(uint16_t sleepcnt);
 void led_ctrl();
 
 /**
@@ -43,7 +44,7 @@ void main(void)
 }
 
 void led_ctrl(){
-    static unsigned int state = 0;
+    static uint8_t state = 0;
     state++;
     switch (state)
     {
@@ -75,8 +76,8 @@ void led_ctrl(){
     }
 }
 
-void sleep(unsigned int sleepcnt){
-    volatile unsigned int i;        // volatile to prevent optimization
+void sleep(uint16_t sleepcnt){
+    volatile uint16_t i;        // volatile to prevent optimization
     for(i=sleepcnt; i>0; i--);      // delay
     return;
 }
